requirements.txt
matplotlib==3.8.3
numpy==1.26.4
pillow==10.2.0
scikit-image==0.22.0
torch==2.2.1
torchvision==0.17.1
grad-cam==1.5.0
requests==2.31.0
streamlit==1.32.2
ipykernel==6.26.0
wandb==0.18.1
fastapi==0.104.1
uvicorn==0.24.0